refactors network_runner.resources to network_runner.models
This commit is a simple package name change from
`network_runner.resources` to `network_runner.models`.  No additional
changes have been made to this commit.
