+++ network_runner/models/inventory.py
