+++ tests/functional.py
@@ -2,7 +2,7 @@ import argparse
 import json
 
 from network_runner import api
-from network_runner.resources.inventory import Inventory
+from network_runner.models.inventory import Inventory
 
 TRUNK_SUPPORT = ('cumulus', 'eos', 'junos', 'openvswitch')
 PORTS = {
