+++ tests/unit/base.py
@@ -17,7 +17,7 @@ import mock
 import unittest
 
 from network_runner import api
-from network_runner.resources.inventory import Inventory
+from network_runner.models.inventory import Inventory
 
 
 class BaseTestCase(unittest.TestCase):
