+++ tests/unit/test_api.py
@@ -16,8 +16,8 @@
 import mock
 
 from network_runner import exceptions
-from network_runner.resources import playbook
-from network_runner.resources.inventory import Host
+from network_runner.models import playbook
+from network_runner.models.inventory import Host
 from network_runner.api import NetworkRunner
 
 from . import base
